코드 리팩토링, MemoryCache 소멸자 설정
